loing and register completed

+++ controller/authController.js
@@ -79,16 +79,19 @@ exports.register = async (req, res, next) => {
       return res.status(400).json({ errPassword: 'password and confirm password did not match' });
     }
     const hasedPassword = await bcrypt.hash(password, 12);
+    let result = null;
 
-    const result = await uploadPromise(req.file.path, { timeout: 60000000 });
+    if (req.file) {
+      result = await uploadPromise(req.file.path, { timeout: 60000000 });
+      fs.unlinkSync(req.file.path);
+    }
     await User.create({
       firstName,
       lastName,
       email,
       password: hasedPassword,
-      profilePicture: result.secure_url,
+      profilePicture: result === null ? null : result.secure_url,
     });
-    fs.unlinkSync(req.file.path);
     res.status(200).json({ message: 'you account has been created' });
   } catch (err) {
     next(err);
